Move asset webpack config to `assets/`.
